Fixes issue#182
