+++ src/app/business/profile/profile.module.ts
@@ -6,7 +6,7 @@ import { ReactiveFormsModule, FormsModule } from '@angular/forms';
 
 import { ProfileCardComponent } from './profileCard/profile-card.component';
 import { ButtonModule,CardModule,ChartModule,MessageModule,OverlayPanelModule,DialogModule ,ConfirmationService,ConfirmDialogModule, FormModule, DropMenuModule} from '../../components/common/api';
-import { ProfileService } from './profile.service';
+import { ProfileService, PoolService } from './profile.service';
 import { HttpService } from '../../shared/api';
 import { SuspensionFrameComponent } from './profileCard/suspension-frame/suspension-frame.component';
 
@@ -39,7 +39,8 @@ let routers = [{
   providers: [
     HttpService,
     ProfileService,
-    ConfirmationService
+    ConfirmationService,
+    PoolService
   ]
 })
 export class ProfileModule { }
